verbose/requirements.txt
asgiref==3.9.1
channels==4.2.2
dj_database_url==3.0.1
Django==5.2.4
django_channels==0.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf_spectacular==0.28.0
Pillow==11.3.0
PyJWT==2.10.1
python-dotenv==1.1.1